+++ CHANGES.md
@@ -9,6 +9,7 @@
 * Add a persons available socials (Youtube, LinkedIn, Reddit, Fansite, TikTok, Wikidata)
 * Change increase viewable history menu items from 13 to 15
 * Change add parsing of 2160p releases that don't have a source tag
+* Change twitter.com to x.com and Twitter to X (Twitter) branding
 
 
 ### 3.32.8 (2024-10-07 00:30:00 UTC)